linux_bionic is linux
linux_bionic is the target which means the cross-compiled (i.e. arm
on x86) host target. libusb is now configured to use linux/config.h for
this new target because it is also considered as Linux. Previously, the
header was used only for the linux_glibc target which means the native
host.

Bug: 159685774
Test: HOST_CROSS_OS=linux_bionic m libusb

Change-Id: I6ad6e58e6c78fc59b958f1430797a19385e7ce41